Multiple changes
EventActionValue is disabled for now since it is not working like I would like to.
ToStringConverterListener renamed to ConvertToEvent and contains basic conversation stuff (int to float, float to int etc).
EventAction (ScriptableObjects) now contains the basic stuff from EventActionValue. Based to this EventActionEditor was created and allows to edit in the inspector the initial values of the SO.
Also EventAction has runtime value which will get assigned from the value of initialvalue. So you can change SO value without to worry.
The name of initial value will be renamed in playmode to runtime value.